refactor(pdpa): decompose PdpaEncryptionService into Status/Backfill/Retention + util
Behavior-preserving (1 $tx data-migration + counter, 0 JE). Facade keeps 8-method
surface + 3-arg ctor + the 4 static consts (DEFAULT_BATCH_SIZE for CLI) + delegates;
internally constructs 3 sub-services (spec + CLI + module untouched). The processBatch
$transaction (writes + progress counter atomic, W5) moved WHOLE; advisory-lock try/finally
+ audit-after-unlock + W9 retry + salvage preserved; never-log-PII intact. tsc 0, pdpa tests green, lint 0.
